requirements.txt
gffutils
deeptools
pyBigWig
numpy
pybedtools
biopython
pandas
sklearn